+++ app/Http/Controllers/UserManagement/UserController.php
@@ -7,6 +7,8 @@ use App\Http\Requests\UserManagement\StoreUser;
 use App\Http\Requests\UserManagement\UpdateUser;
 use App\Models\Role;
 use App\Models\User;
+use App\Util\AutoColorInitialAvatar;
+use App\View\Components\UserAvatar;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Str;
@@ -223,4 +225,28 @@ class UserController extends Controller
             ->with('success', __('userprofile.oauth_disabled_new_password_has_been_set'));
     }
 
+    /**
+     * Display the avatar of the user.
+     *
+     * @param \App\Models\User $user
+     * @return \Illuminate\Http\Response
+     */
+    public function avatar(User $user, Request $request)
+    {
+        $request->validate([
+            'size' => [
+                'nullable',
+                'integer',
+                'min:10',
+                'max:250',
+            ],
+        ]);
+
+        $avatar = new AutoColorInitialAvatar();
+        return $avatar->name($user->name)
+            ->size($request->input('size', UserAvatar::DEFAULT_SIZE))
+            ->autoColor()
+            ->generate()
+            ->response('png', 100);
+    }
 }