+++ resources/js/components/UserAvatar.vue
@@ -1,0 +1,22 @@
+<template>
+    <img
+        :src="url"
+        class="avatar"
+        :width="size"
+        :height="size"
+    />
+</template>
+
+<script>
+export default {
+    props: {
+        url: {
+            required: true
+        },
+        size: {
+            required: false,
+            default: 48,
+        }
+    }
+}
+</script>