+++ resources/js/pages/users/UserIndexPage.vue
@@ -7,12 +7,11 @@
         :empty-text="$t('app.no_users_found')"
         :items-per-page="25"
     >
-        <template v-slot:cell(avatar)="data">
-            <img
-                :src="data.item.avatar_url_site_header"
-                alt="Gravatar"
-                style="width: 30px; height: 30px;"
-            >
+        <template v-slot:cell(avatar_url)="data">
+            <user-avatar
+                :url="data.value"
+                size="30"
+            />
         </template>
         <template v-slot:cell(name)="data">
             <a :href="data.item.links.show">
@@ -54,17 +53,19 @@
 import moment from 'moment'
 import BaseTable from '@/components/table/BaseTable'
 import EmailLink from '@/components/common/EmailLink'
+import UserAvatar from '@/components/UserAvatar'
 import usersApi from '@/api/user_management/users'
 export default {
     components: {
         BaseTable,
-        EmailLink
+        EmailLink,
+        UserAvatar
     },
     data() {
         return {
             fields: [
                 {
-                    key: 'avatar',
+                    key: 'avatar_url',
                     label: this.$t('userprofile.avatar'),
                     class: 'fit align-middle text-center'
                 },