+++ resources/views/components/user-avatar.blade.php
@@ -1,0 +1,6 @@
+<img
+    src="{{ $user->avatarUrl($size) }}"
+    class="avatar"
+    width="{{ $size }}"
+    height="{{ $size }}"
+/>